Ain't pretty, but it seems to work.
I'm gonna leave it at that. It's been a rough night. My secondary monitor doesn't work. How am I supposed to code like this. :)
