+++ views/cars/create.hbs
@@ -1,4 +1,4 @@
-<form action="/api/cars" method="POST" enctype="multipart/form-data" 
+<form action="/cars" method="POST" enctype="multipart/form-data" 
       class="flex flex-col p-2 col-6 w-4/5 justify-center items-center mx-auto"
       >
   <div class="flex p-2 w-3/5 justify-between items-center">
