+++ views/cars/show.hbs
@@ -2,7 +2,7 @@
 <h2>Posted by {{car.user.username}}</h2>
 <h4>Mileage: {{car.mileage}}</h4>
 <br>
-<img src="/images/{{car.filename}}">
+<img src="{{car.file_path}}">
 <p>{{car.description}}</p>
 <br>
